requirements.txt
pandas
click
scipy==1.4.1
numpy==1.18.1
Pillow==8.0.1 
scikit_learn==0.23.2 
networkx==2.5 
gensim==3.6.0 
tensorflow==1.15.0 
python_igraph==0.8.3 
Click==7.0